# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateStoryRequest:
    properties:
      author_name:
        type: string
      cover_image_url:
        type: string
      description:
        type: string
      genre_ids:
        items:
          type: string
        type: array
      is_published:
        type: boolean
      status:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Pagination:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API cho nền tảng đọc truyện web novel Nekozanedex
  title: Nekozanedex API
  version: "1.0"
paths:
...
  /api/admin/stories:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Pagination'
      security:
      - BearerAuth: []
      summary: Lấy tất cả truyện (Admin)
      tags:
      - Admin - Stories
    post:
      consumes:
      - application/json
      parameters:
      - description: Story Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateStoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tạo truyện mới (Admin)
      tags:
      - Admin - Stories
  /api/admin/stories/{id}:
    delete:
      parameters:
      - description: Story ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Xóa truyện (Admin)
      tags:
      - Admin - Stories
    get:
      parameters:
      - description: Story ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lấy truyện theo ID (Admin)
      tags:
      - Admin - Stories
    put:
      consumes:
      - application/json
      parameters:
      - description: Story ID
        in: path
        name: id
        required: true
        type: string
      - description: Story Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateStoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Cập nhật truyện (Admin)
      tags:
      - Admin - Stories
...
  /api/genres/{genre}/stories:
    get:
      parameters:
      - description: Genre Slug
        in: path
        name: genre
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Pagination'
      summary: Lấy truyện theo thể loại
      tags:
      - Stories
  /api/notifications:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Pagination'
      security:
      - BearerAuth: []
      summary: Lấy danh sách thông báo của tôi
      tags:
      - Notifications
  /api/notifications/{id}/read:
    post:
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Đánh dấu đã đọc
      tags:
      - Notifications
  /api/notifications/read-all:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Đánh dấu tất cả đã đọc
      tags:
      - Notifications
  /api/notifications/unread-count:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lấy số thông báo chưa đọc
      tags:
      - Notifications
  /api/stories:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Pagination'
      summary: Lấy danh sách truyện
      tags:
      - Stories
  /api/stories/{slug}:
    get:
      parameters:
      - description: Story Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Lấy chi tiết truyện theo slug
      tags:
      - Stories
...
  /api/stories/hot:
    get:
      parameters:
      - default: 10
        description: Number of stories
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Lấy truyện hot
      tags:
      - Stories
  /api/stories/latest:
    get:
      parameters:
      - default: 10
        description: Number of stories
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Lấy truyện mới cập nhật
      tags:
      - Stories
...
  /api/stories/search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Pagination'
      summary: Tìm kiếm truyện
      tags:
      - Stories
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"